Switched some render_template to redirect
It's important to differentiate between these two methods. If you're in the current page, and you return to the same page, then you should use render. But if you go to a different page while using render - it'll actually say that you're still in the same page, just that it looks like something else.